backend/requirements.txt
base58
firebase-admin
Flask==1.1.2
flask-cors==3.0.8
google-auth
google-api-python-client
google-auth-httplib2
google-cloud-pubsub
pyjwt==1.7.1
requests==2.23.0
requests-toolbelt==0.9.1
werkzeug==1.0.1
